Mergestream fixes by feedback
