+++ lib/MergeStream.js
@@ -1,8 +1,7 @@
 var Readable = require('stream').Readable,
     Transform = require('stream').Transform,
     util = require('util'),
-    moment = require('moment'),
-    through2 = require('through2');
+    AddStreamNameTransformer = require('./AddStreamNameTransformer');
 
 /** A merger returns connections from different endpoints while maintaining departure time ordered. */
 var MergeStream = function(connectionsStreams, departureTimeSync) {
@@ -15,7 +14,6 @@ var MergeStream = function(connectionsStreams, departureTimeSync) {
   this._departureTime = departureTimeSync; // Keeps track of departureTime that we return connections from
   this._maxInQueues = departureTimeSync; // Maximum departure time in queue
   this._amountConnectionsInQueues = {}; // Counts amount of connections inside queues per stream
-  this._streamActive = {}; // holds bool whether a stream is being processed or not
 
   for (var k=0; k<connectionsStreams.length; k++) {
     this.addConnectionsStream(connectionsStreams[k]);
@@ -35,23 +33,18 @@ MergeStream.prototype.addConnectionsStream = function (connectionsStream) {
   this._connectionsStreamIndex[streamName] = index; // map name - index stream
   this._amountConnectionsInQueues[streamName] = 0;
   this._amountStreams++;
-  this._streamActive[streamName] = true;
 
   this._connectionsStreams[index] = connectionsStream[1]
-                                    .pipe(new AddStreamNameTransformer(streamName))
-                                    .pipe(through2.obj(function (connection, enc, done) {
-                                      // Don't process departed connection
-                                      if ((connection['@context'] || connection['departureTime'] >= self._departureTime) && self._streamActive[streamName]) {
-                                        this.push(connection);
-                                      }
-                                      done();
-                                    }));
+                                    .pipe(new AddStreamNameTransformer(streamName));
 
   // Add event listener to stream
   this._connectionsStreams[index].on('data', function (connection) {
-    // Count available connections per stream
-    self._amountConnectionsInQueues[connection['streamName']]++;
-    self._connectionListener(connection);
+    // Don't process context or departed connection
+    if (!connection['@context'] && connection['departureTime'] >= self._departureTime) {
+      // Count available connections per stream
+      self._amountConnectionsInQueues[connection['streamName']]++;
+      self._connectionListener(connection);
+    }
   });
 
   this._connectionsStreams[index].on('end', function() {
@@ -65,7 +58,9 @@ MergeStream.prototype.addConnectionsStream = function (connectionsStream) {
   });
 };
 
-MergeStream.prototype.removeConnectionsStream = function (connectionsStreamName) {
+// First parameter: name of the stream you want to stop
+// Second parameter: callback that ends the stream
+MergeStream.prototype.stopConnectionsStream = function (connectionsStreamName, endConnectionStream) {
   this._amountStreams--;
   // Close mergestream if last available stream
   if (this._amountStreams == 0) {
@@ -73,20 +68,10 @@ MergeStream.prototype.removeConnectionsStream = function (connectionsStreamName)
   } else {
     var index = this._connectionsStreamIndex[connectionsStreamName];
     if (this._connectionsStreams[index]) {
-      this._connectionsStreams[index].destroy();
+      endConnectionStream();
       this._connectionsStreams[index] = null;
     }
   }
-};
-
-// Starts merging stream again
-MergeStream.prototype.startConnectionsStream = function (connectionsStreamName) {
-  this._streamActive[connectionsStreamName] = true;
-};
-
-// Stop merging of stream
-MergeStream.prototype.stopConnectionsStream = function (connectionsStreamName) {
-  this._streamActive[connectionsStreamName] = false;
 };
 
 // Returns all connections that depart
@@ -116,26 +101,15 @@ MergeStream.prototype._read = function() {
 
 // Returns minimum connection from queues
 MergeStream.prototype._next = function() {
-  // Get context if available
-  var context = this._getContextIfAvailable();
-  if (context != null) {
-    return context;
-  }
-
   var minimumTime = null;
 
   var i = 0;
   while (i <= this._queues.length) {
     if (this._queues[i] && this._queues[i].length > 0) {
-      if (this._queues[i][0]['@context'] || this._queues[i][0]['departureTime'].getTime() == this._departureTime.getTime()) {
+      if (this._queues[i][0]['departureTime'].getTime() == this._departureTime.getTime()) {
         var connection = this._queues[i].shift();
         this._amountConnectionsInQueues[connection['streamName']] -= 1;
-        // Only return connection when the stream hasn't been stopped
-        if (this._streamActive[connection['streamName']]) {
           return connection;
-        } else {
-          return this._next();
-        }
       } else if (this._queues[i] && (minimumTime == null || this._queues[i][0]['departureTime'] < minimumTime)) {
         minimumTime = this._queues[i][0]['departureTime']; // New minimum departure time
       }
@@ -166,7 +140,7 @@ MergeStream.prototype.close = function () {
   // Close all connectionsStreams
   for (var k=0; k<this._connectionsStreams.length; k++) {
     if (this._connectionsStreams[k]) {
-      this._connectionsStreams[k].destroy();
+      this._connectionsStreams[k] = null;
     }
   }
 
@@ -176,7 +150,7 @@ MergeStream.prototype.close = function () {
 MergeStream.prototype._connectionListener = function(connection) {
   // Load connections until connection is bigger then latest departure time
 
-  if (!connection['@context'] && connection['departureTime'] > this._maxInQueues) {
+  if (connection['departureTime'] > this._maxInQueues) {
     // Pause the stream
     var streamName = connection['streamName'];
     var streamIndex = this._connectionsStreamIndex[streamName];
@@ -196,19 +170,16 @@ MergeStream.prototype._connectionListener = function(connection) {
 MergeStream.prototype._addConnection = function (connection) {
   var added = false;
 
-  // If not a context
-  if (!connection['@context']) {
-    var i = 0;
-    while (!added && i<this._queues.length) {
-      var length = this._queues[i].length;
+  var i = 0;
+  while (!added && i<this._queues.length) {
+    var length = this._queues[i].length;
 
-      // If the queue is empty or connection departureTime is later then last connection in the queue
-      if (length == 0 || this._queues[i][length - 1]['departureTime'] <= connection['departureTime']) {
-        this._queues[i].push(connection);
-        added = true;
-      }
-      i++;
+    // If the queue is empty or connection departureTime is later then last connection in the queue
+    if (length == 0 || this._queues[i][length - 1]['departureTime'] <= connection['departureTime']) {
+      this._queues[i].push(connection);
+      added = true;
     }
+    i++;
   }
 
   if (!added) {
@@ -220,17 +191,6 @@ MergeStream.prototype._addConnection = function (connection) {
   // Update maximum departure time in queue
   if (connection['departureTime'] > this._maxInQueues) {
     this._maxInQueues = connection['departureTime'];
-  }
-};
-
-
-MergeStream.prototype._getContextIfAvailable = function() {
-  var i = 0;
-  while (i <= this._queues.length) {
-    if (this._queues[i] && this._queues[i].length > 0 && this._queues[i][0]['@context']) {
-      return this._queues[i].shift();
-    }
-    i++;
   }
 };
 
@@ -260,21 +220,5 @@ MergeStream.prototype._allActiveStreamsAvailable = function() {
   }
   return true;
 }
-
-/**
- * Add name of stream to connections
- */
-function AddStreamNameTransformer (name) {
-  Transform.call(this, {objectMode : true});
-  this._name = name;
-}
-
-util.inherits(AddStreamNameTransformer, Transform);
-
-AddStreamNameTransformer.prototype._transform = function (connection, encoding, done) {
-  connection['streamName'] = this._name;
-  this.push(connection);
-  done();
-};
 
 module.exports = MergeStream;
