+++ package.json
@@ -25,10 +25,8 @@
   "devDependencies": {
     "async": "^1.2.1",
     "mocha": "^2.3.3",
-    "moment": "^2.10.6",
     "should": "^6.0.3"
   },
   "dependencies": {
-    "through2": "^2.0.0"
   }
 }
