+++ test/data/connections-ns.jsonldstream
@@ -1,4 +1,3 @@
-{"@context":{"gtfs":"http://vocab.gtfs.org/terms#","trip":"gtfs:trip","route":"gtfs:route","headsign":"gtfs:headsign","lc":"http://semweb.mmlab.be/ns/linkedconnections#","arrivalStop":"lc:arrivalStop","arrivalTime":"lc:arrivalTime","departureStop":"lc:departureStop","departureTime":"lc:departureTime","Connection":"lc:Connection","connection":"http://linkedconnections.org/connections/"}}
 {"@type":"Connection","@id":"connection:483631","arrivalTime":"2015-10-10T09:54:00.000Z","arrivalStop":"hlo","departureTime":"2015-10-10T09:49:00.000Z","departureStop":"amr","trip":"10435|11|3439","route":"100-IC","headsign":"Haarlem"}
 {"@type":"Connection","@id":"connection:483632","arrivalTime":"2015-10-10T09:59:00.000Z","arrivalStop":"cas","departureTime":"2015-10-10T09:54:00.000Z","departureStop":"hlo","trip":"10435|11|3439","route":"100-IC","headsign":"Haarlem"}
 {"@type":"Connection","@id":"connection:483633","arrivalTime":"2015-10-10T10:04:00.000Z","arrivalStop":"utg","departureTime":"2015-10-10T09:59:00.000Z","departureStop":"cas","trip":"10435|11|3439","route":"100-IC","headsign":"Haarlem"}
